add move_goal

+++ catkin_ws/src/navigation/src/record.py
@@ -3,10 +3,11 @@ import wave
 
 CHUNK = 1024
 FORMAT = pyaudio.paInt16
-CHANNELS = 2 # モノラル入力
-RATE = 44100
-RECORD_SECONDS = 5
-WAVE_OUTPUT_FILENAME = "output.wav"
+CHANNELS = 2
+RATE = 16000 # Can kinect deal with 44100Hz?
+RECORD_SECONDS = 20 # record time is accurate?
+SAMPLE_SECONDS = 5
+index = 4
 
 p = pyaudio.PyAudio()
 
@@ -14,25 +15,36 @@ stream = p.open(format=FORMAT,
                 channels=CHANNELS,
                 rate=RATE,
                 input=True,
+                input_device_index=index,
                 frames_per_buffer=CHUNK)
 
 print("* recording")
 
+"""
 frames = []
-
 for i in range(0, int(RATE / CHUNK * RECORD_SECONDS)):
+    frame = []
     data = stream.read(CHUNK)
     frames.append(data)
+    if len(frames) == RATE / CHUNK * SAMPLE_SECONDS:
+"""        
 
-print("* done recording")
 
-stream.stop_stream()
-stream.close()
-p.terminate()
+for i in range(0, int(RECORD_SECONDS / SAMPLE_SECONDS)):
+    frames = []
+    WAVE_OUTPUT_FILENAME = "output" + str(i) + ".wav"
+    for j in range(0, int(RATE / CHUNK * SAMPLE_SECONDS)):
+        data = stream.read(CHUNK)
+        frames.append(data)
+    print("* done recording")
+    if i == (RECORD_SECONDS / SAMPLE_SECONDS):
+        stream.stop_stream()
+        stream.close()
+        p.terminate()
 
-wf = wave.open("./wavfile/" + WAVE_OUTPUT_FILENAME, 'wb')
-wf.setnchannels(CHANNELS)
-wf.setsampwidth(p.get_sample_size(FORMAT))
-wf.setframerate(RATE)
-wf.writeframes(b''.join(frames))
-wf.close()
+    wf = wave.open("./wavfile/" + WAVE_OUTPUT_FILENAME, "wb")
+    wf.setnchannels(CHANNELS)
+    wf.setsampwidth(p.get_sample_size(FORMAT))
+    wf.setframerate(RATE)
+    wf.writeframes(b''.join(frames))
+    wf.close()
